move overlapmode tests
