+++ src/symbol/symbol_style_layer.test.ts
@@ -1,5 +1,4 @@
 import SymbolStyleLayer from '../style/style_layer/symbol_style_layer';
-import {getOverlapMode} from '../style/style_layer/overlap_mode';
 import FormatSectionOverride from '../style/format_section_override';
 import properties, {SymbolPaintPropsPossiblyEvaluated} from '../style/style_layer/symbol_style_layer_properties.g';
 import ZoomHistory from '../style/zoom_history';
@@ -101,51 +100,4 @@ describe('hasPaintOverrides', () => {
 
     });
 
-});
-
-describe('getOverlapMode', () => {
-    test('defaults - no props set', () => {
-        const props = {};
-        const layer = createSymbolLayer(props);
-
-        expect(getOverlapMode(layer.layout, 'icon-overlap', 'icon-allow-overlap')).toBe('never');
-        expect(getOverlapMode(layer.layout, 'text-overlap', 'text-allow-overlap')).toBe('never');
-    });
-
-    test('-allow-overlap set', () => {
-        const props = {layout: {'icon-allow-overlap': false, 'text-allow-overlap': true}};
-        const layer = createSymbolLayer(props);
-
-        expect(getOverlapMode(layer.layout, 'icon-overlap', 'icon-allow-overlap')).toBe('never');
-        expect(getOverlapMode(layer.layout, 'text-overlap', 'text-allow-overlap')).toBe('always');
-    });
-
-    test('-overlap set', () => {
-        let props = {layout: {'icon-overlap': 'never', 'text-overlap': 'always'}};
-        let layer = createSymbolLayer(props);
-
-        expect(getOverlapMode(layer.layout, 'icon-overlap', 'icon-allow-overlap')).toBe('never');
-        expect(getOverlapMode(layer.layout, 'text-overlap', 'text-allow-overlap')).toBe('always');
-
-        props = {layout: {'icon-overlap': 'always', 'text-overlap': 'cooperative'}};
-        layer = createSymbolLayer(props);
-
-        expect(getOverlapMode(layer.layout, 'icon-overlap', 'icon-allow-overlap')).toBe('always');
-        expect(getOverlapMode(layer.layout, 'text-overlap', 'text-allow-overlap')).toBe('cooperative');
-    });
-
-    test('-overlap beats -allow-overlap', () => {
-        const props = {
-            layout: {
-                'icon-overlap': 'never',
-                'icon-allow-overlap': true,
-                'text-overlap': 'cooperative',
-                'text-allow-overlap': false
-            }
-        };
-        const layer = createSymbolLayer(props);
-
-        expect(getOverlapMode(layer.layout, 'icon-overlap', 'icon-allow-overlap')).toBe('never');
-        expect(getOverlapMode(layer.layout, 'text-overlap', 'text-allow-overlap')).toBe('cooperative');
-    });
 });
